unify query string building logic
